Crate zcash_address

Modules

unified

Structs

UnsupportedAddress

An address type is not supported for conversion.

ZcashAddress

A Zcash address.

Enums

Network

The Zcash network for which an address is encoded.

ParseError

An error while attempting to parse a string as a Zcash address.

Traits

FromAddress

A helper trait for converting a ZcashAddress into another type.

ToAddress

A helper trait for converting another type into a ZcashAddress.