Struct zcash_address::ZcashAddress

pub struct ZcashAddress { /* private fields */ }

A Zcash address.

Implementations

impl ZcashAddress

pub fn try_from_encoded(s: &str) -> Result<Self, ParseError>

Attempts to parse the given string as a Zcash address.

This simply calls s.parse(), leveraging the FromStr implementation.

Errors

If the parser can detect that the string must contain an address encoding used by Zcash, ParseError::InvalidEncoding will be returned if any subsequent part of that encoding is invalid.

In all other cases, ParseError::NotZcash will be returned on failure.

Examples

use zcash_address::ZcashAddress;

let encoded = "zs1z7rejlpsa98s2rrrfkwmaxu53e4ue0ulcrw0h4x5g8jl04tak0d3mm47vdtahatqrlkngh9sly";
let addr = ZcashAddress::try_from_encoded(&encoded);
assert_eq!(encoded.parse(), addr);

pub fn convert<T: FromAddress>(self) -> Result<T, UnsupportedAddress>

Converts this address into another type.

convert can convert into any type that implements the FromAddress trait. This enables ZcashAddress to be used as a common parsing and serialization interface for Zcash addresses, while delegating operations on those addresses (such as constructing transactions) to downstream crates.

If you want to get the encoded string for this address, use the Display implementation instead via address.to_string().