zcash_address

Module unified

Implementation of ZIP 316 Unified Addresses and Viewing Keys.

Structs

Address

A Unified Address.

Ufvk

A Unified Full Viewing Key.

Uivk

A Unified Incoming Viewing Key.

Enums

Bech32mZip316

The bech32m checksum algorithm, defined in BIP-350, extended to allow all lengths supported by ZIP 316.

Fvk

The set of known FVKs for Unified FVKs.

Ivk

The set of known IVKs for Unified IVKs.

ParseError

An error while attempting to parse a string as a Zcash address.

Receiver

The set of known Receivers for Unified Addresses.

Typecode

The known Receiver and Viewing Key types.

Traits

Container

Trait for Unified containers, that exposes the items within them.

Encoding

Trait providing common encoding and decoding logic for Unified containers.