zcash_address::unified

Trait Container

pub trait Container {
    type Item: SealedItem;

    // Required method
    fn items_as_parsed(&self) -> &[Self::Item];

    // Provided method
    fn items(&self) -> Vec<Self::Item> { ... }
}

Trait for Unified containers, that exposes the items within them.

Required Associated Types

type Item: SealedItem

The type of item in this unified container.

Required Methods

fn items_as_parsed(&self) -> &[Self::Item]

Returns the items in the order they were parsed from the string encoding.

This API is for advanced usage; in most cases you should use Self::items.

Provided Methods

fn items(&self) -> Vec<Self::Item>

Returns the items contained within this container, sorted in preference order.